3.1 Integer overflow in area calculations

+++ src/slideio/base/size.hpp
@@ -51,8 +51,8 @@ namespace slideio
             return width == other.width && height == other.height;
         }
 
-        int32_t area() const {
-            return width * height;
+        int64_t area() const {
+            return static_cast<int64_t>(width) * height;
         }
         bool empty() const {
             return width <= 0 || height <= 0;
